+++ src/RootCloseWrapper.js
@@ -1,14 +1,12 @@
+import contains from 'dom-helpers/query/contains';
+import PropTypes from 'prop-types';
 import React from 'react';
 import ReactDOM from 'react-dom';
 
 import addEventListener from './utils/addEventListener';
-import createChainedFunction from './utils/createChainedFunction';
 import ownerDocument from './utils/ownerDocument';
 
-// TODO: Consider using an ES6 symbol here, once we use babel-runtime.
-const CLICK_WAS_INSIDE = '__click_was_inside';
-
-let counter = 0;
+const escapeKeyCode = 27;
 
 function isLeftClickEvent(event) {
   return event.button === 0;
@@ -18,134 +16,118 @@ function isModifiedEvent(event) {
   return !!(event.metaKey || event.altKey || event.ctrlKey || event.shiftKey);
 }
 
-function getSuppressRootClose() {
-  let id = CLICK_WAS_INSIDE + '_' + counter++;
-  return {
-    id,
-    suppressRootClose(event) {
-      // Tag the native event to prevent the root close logic on document click.
-      // This seems safer than using event.nativeEvent.stopImmediatePropagation(),
-      // which is only supported in IE >= 9.
-      event.nativeEvent[id] = true;
-    }
-  };
-}
+/**
+ * The `<RootCloseWrapper/>` component registers your callback on the document
+ * when rendered. Powers the `<Overlay/>` component. This is used achieve modal
+ * style behavior where your callback is triggered when the user tries to
+ * interact with the rest of the document or hits the `esc` key.
+ */
+class RootCloseWrapper extends React.Component {
+  constructor(props, context) {
+    super(props, context);
 
-export default class RootCloseWrapper extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.handleDocumentMouse = this.handleDocumentMouse.bind(this);
-    this.handleDocumentKeyUp = this.handleDocumentKeyUp.bind(this);
-
-    let { id, suppressRootClose } = getSuppressRootClose();
-
-    this._suppressRootId = id;
-
-    this._suppressRootCloseHandler = suppressRootClose;
-  }
-
-  bindRootCloseHandlers() {
-    const doc = ownerDocument(this);
-
-    this._onDocumentMouseListener =
-      addEventListener(doc, this.props.event, this.handleDocumentMouse);
-
-    this._onDocumentKeyupListener =
-      addEventListener(doc, 'keyup', this.handleDocumentKeyUp);
-  }
-
-  handleDocumentMouse(e) {
-    // This is now the native event.
-    if (e[this._suppressRootId]) {
-      return;
-    }
-
-    if (this.props.disabled || isModifiedEvent(e) || !isLeftClickEvent(e)) {
-      return;
-    }
-
-    this.props.onRootClose &&
-      this.props.onRootClose();
-  }
-
-  handleDocumentKeyUp(e) {
-    if (e.keyCode === 27 && this.props.onRootClose) {
-      this.props.onRootClose();
-    }
-  }
-
-  unbindRootCloseHandlers() {
-    if (this._onDocumentMouseListener) {
-      this._onDocumentMouseListener.remove();
-    }
-
-    if (this._onDocumentKeyupListener) {
-      this._onDocumentKeyupListener.remove();
-    }
+    this.preventMouseRootClose = false;
   }
 
   componentDidMount() {
-    this.bindRootCloseHandlers();
-  }
-
-  render() {
-    const {event, noWrap, children} = this.props;
-    const child = React.Children.only(children);
-
-    const handlerName = event == 'click' ? 'onClick' : 'onMouseDown';
-
-    if (noWrap) {
-      return React.cloneElement(child, {
-        [handlerName]: createChainedFunction(this._suppressRootCloseHandler, child.props[handlerName])
-      });
+    if (!this.props.disabled) {
+      this.addEventListeners();
     }
-
-    // Wrap the child in a new element, so the child won't have to handle
-    // potentially combining multiple onClick listeners.
-    return (
-      <div {...{[handlerName]: this._suppressRootCloseHandler}}>
-        {child}
-      </div>
-    );
   }
 
-  getWrappedDOMNode() {
-    // We can't use a ref to identify the wrapped child, since we might be
-    // stealing the ref from the owner, but we know exactly the DOM structure
-    // that will be rendered, so we can just do this to get the child's DOM
-    // node for doing size calculations in OverlayMixin.
-    const node = ReactDOM.findDOMNode(this);
-    return this.props.noWrap ? node : node.firstChild;
+  componentDidUpdate(prevProps) {
+    if (!this.props.disabled && prevProps.disabled) {
+      this.addEventListeners();
+    } else if (this.props.disabled && !prevProps.disabled) {
+      this.removeEventListeners();
+    }
   }
 
   componentWillUnmount() {
-    this.unbindRootCloseHandlers();
+    if (!this.props.disabled) {
+      this.removeEventListeners();
+    }
+  }
+
+  addEventListeners = () => {
+    const { event } = this.props;
+    const doc = ownerDocument(this);
+
+    // Use capture for this listener so it fires before React's listener, to
+    // avoid false positives in the contains() check below if the target DOM
+    // element is removed in the React mouse callback.
+    this.documentMouseCaptureListener =
+      addEventListener(doc, event, this.handleMouseCapture, true);
+
+    this.documentMouseListener =
+      addEventListener(doc, event, this.handleMouse);
+
+    this.documentKeyupListener =
+      addEventListener(doc, 'keyup', this.handleKeyUp);
+  }
+
+  removeEventListeners = () => {
+    if (this.documentMouseCaptureListener) {
+      this.documentMouseCaptureListener.remove();
+    }
+
+    if (this.documentMouseListener) {
+      this.documentMouseListener.remove();
+    }
+
+    if (this.documentKeyupListener) {
+      this.documentKeyupListener.remove();
+    }
+  }
+
+  handleMouseCapture = (e) => {
+    this.preventMouseRootClose = (
+      isModifiedEvent(e) ||
+      !isLeftClickEvent(e) ||
+      contains(ReactDOM.findDOMNode(this), e.target)
+    );
+  };
+
+  handleMouse = (e) => {
+    if (!this.preventMouseRootClose && this.props.onRootClose) {
+      this.props.onRootClose(e);
+    }
+  };
+
+  handleKeyUp = (e) => {
+    if (e.keyCode === escapeKeyCode && this.props.onRootClose) {
+      this.props.onRootClose(e);
+    }
+  };
+
+  render() {
+    return this.props.children;
   }
 }
 
 RootCloseWrapper.displayName = 'RootCloseWrapper';
 
 RootCloseWrapper.propTypes = {
-  onRootClose: React.PropTypes.func,
-
   /**
-   * Disable the the RootCloseWrapper, preventing it from triggering
-   * `onRootClose`.
+   * Callback fired after click or mousedown. Also triggers when user hits `esc`.
    */
-  disabled: React.PropTypes.bool,
+  onRootClose: PropTypes.func,
   /**
-   * Passes the suppress click handler directly to the child component instead
-   * of placing it on a wrapping div. Only use when you can be sure the child
-   * properly handle the click event.
+   * Children to render.
    */
-  noWrap: React.PropTypes.bool,
+  children: PropTypes.element,
   /**
-   * Choose which document mouse event to bind to
+   * Disable the the RootCloseWrapper, preventing it from triggering `onRootClose`.
    */
-  event: React.PropTypes.oneOf(['click', 'mousedown'])
+  disabled: PropTypes.bool,
+  /**
+   * Choose which document mouse event to bind to.
+   */
+  event: PropTypes.oneOf(['click', 'mousedown'])
 };
 
 RootCloseWrapper.defaultProps = {
   event: 'click'
 };
+
+export default RootCloseWrapper;