+++ test/ModalSpec.js
@@ -1,11 +1,10 @@
 import jQuery from 'jquery';
 import React from 'react';
-import ReactTestUtils from 'react-addons-test-utils';
 import ReactDOM from 'react-dom';
+import ReactTestUtils from 'react-dom/test-utils';
 import simulant from 'simulant';
 
 import Modal from '../src/Modal';
-
 import Transition from '../src/Transition';
 
 import { render, shouldWarn } from './helpers';
@@ -21,8 +20,10 @@ describe('Modal', function () {
   });
 
   afterEach(function () {
-    ReactDOM.unmountComponentAtNode(mountPoint);
-    document.body.removeChild(mountPoint);
+    const unmounted = ReactDOM.unmountComponentAtNode(mountPoint);
+    if (unmounted) {
+      document.body.removeChild(mountPoint);
+    }
   });
 
   it('Should render the modal content', function() {
@@ -32,17 +33,15 @@ describe('Modal', function () {
       </Modal>
     , mountPoint);
 
-    assert.equal(instance.refs.modal.querySelectorAll('strong').length, 1);
+    expect(instance.modalNode.querySelectorAll('strong').length).to.equal(1);
   });
 
   it('Should disable scrolling on the modal container while open', function() {
-    let Container = React.createClass({
-      getInitialState() {
-        return { modalOpen: true };
-      },
-      handleCloseModal() {
+    class Container extends React.Component {
+      state = { modalOpen: true };
+      handleCloseModal = () => {
         this.setState({ modalOpen: false });
-      },
+      }
       render() {
         return (
           <div>
@@ -56,11 +55,11 @@ describe('Modal', function () {
           </div>
         );
       }
-    });
+    }
 
     let instance = render(<Container />, mountPoint);
     let modal = ReactTestUtils.findRenderedComponentWithType(instance, Modal);
-    let backdrop = modal.refs.backdrop;
+    let backdrop = modal.backdrop;
 
     expect($(instance).css('overflow')).to.equal('hidden');
 
@@ -70,13 +69,11 @@ describe('Modal', function () {
   });
 
   it('Should add and remove container classes', function() {
-    let Container = React.createClass({
-      getInitialState() {
-        return { modalOpen: true };
-      },
-      handleCloseModal() {
+    class Container extends React.Component {
+      state = { modalOpen: true };
+      handleCloseModal = () => {
         this.setState({ modalOpen: false });
-      },
+      }
       render() {
         return (
           <div>
@@ -91,11 +88,11 @@ describe('Modal', function () {
           </div>
         );
       }
-    });
+    }
 
     let instance = render(<Container />, mountPoint);
     let modal = ReactTestUtils.findRenderedComponentWithType(instance, Modal);
-    let backdrop = modal.refs.backdrop;
+    let backdrop = modal.backdrop;
 
     expect($(instance).hasClass('test test2')).to.be.true;
 
@@ -112,7 +109,7 @@ describe('Modal', function () {
       </Modal>
     , mountPoint);
 
-    let backdrop = instance.refs.backdrop;
+    let backdrop = instance.backdrop;
 
     ReactTestUtils.Simulate.click(backdrop);
 
@@ -127,7 +124,7 @@ describe('Modal', function () {
       </Modal>
     , mountPoint);
 
-    let backdrop = instance.refs.backdrop;
+    let backdrop = instance.backdrop;
 
     ReactTestUtils.Simulate.click(backdrop);
   });
@@ -140,7 +137,7 @@ describe('Modal', function () {
       </Modal>
     , mountPoint);
 
-    let backdrop = instance.refs.backdrop;
+    let backdrop = instance.backdrop;
 
     ReactTestUtils.Simulate.click(backdrop);
 
@@ -155,7 +152,7 @@ describe('Modal', function () {
       </Modal>
     , mountPoint);
 
-    let backdrop = instance.refs.backdrop;
+    let backdrop = instance.backdrop;
 
     simulant.fire(backdrop, 'keyup', { keyCode: 27 });
   });
@@ -164,13 +161,12 @@ describe('Modal', function () {
   it('Should set backdrop Style', function () {
 
     let instance = render(
-      <Modal show bsClass='mymodal' backdropStyle={{ borderWidth: '3px' }}>
+      <Modal show className='mymodal' backdrop backdropStyle={{ borderWidth: '3px' }}>
         <strong>Message</strong>
       </Modal>
     , mountPoint);
 
-    let backdrop = instance.refs.backdrop;
-
+    let backdrop = instance.backdrop;
     expect(
       backdrop.style.borderWidth).to.equal('3px');
   });
@@ -183,7 +179,7 @@ describe('Modal', function () {
           <strong>Message</strong>
         </Modal>
       , mountPoint);
-    }).to.throw(/onlyChild must be passed a children with exactly one child/);
+    }).to.throw(/React.Children.only expected to receive a single React element child./);
   });
 
   it('Should add role to child', function () {
@@ -287,6 +283,29 @@ describe('Modal', function () {
     expect(onShowSpy).to.have.been.calledOnce;
   });
 
+  it('Should accept role on the Modal', function () {
+    let instance = render(
+      <Modal role="alertdialog" show>
+        <strong>Message</strong>
+      </Modal>
+    , mountPoint);
+
+    let attr = instance.modalNode.attributes.getNamedItem('role').value;
+    expect(attr).to.equal('alertdialog');
+  });
+
+  it('Should accept the `aria-describedby` property on the Modal', function () {
+
+    let instance = render(
+      <Modal aria-describedby="modal-description" show>
+        <strong id="modal-description">Message</strong>
+      </Modal>
+    , mountPoint);
+
+    let attr = instance.modalNode.attributes.getNamedItem('aria-describedby').value;
+    expect(attr).to.equal('modal-description');
+  });
+
   describe('Focused state', function () {
     let focusableContainer = null;
 
@@ -348,8 +367,7 @@ describe('Modal', function () {
       document.activeElement.should.equal(input);
     });
 
-    it('Should return focus to the modal', function () {
-
+    it('Should return focus to the modal', () => {
       document.activeElement.should.equal(focusableContainer);
 
       render(
@@ -360,7 +378,9 @@ describe('Modal', function () {
         </Modal>
         , focusableContainer);
 
+
       focusableContainer.focus();
+
       document.activeElement.className.should.contain('modal');
     });